backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
